+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Add json output of tipsets to `louts chain list`. ([filecoin-project/lotus#12691](https://github.com/filecoin-project/lotus/pull/12691))
 - Remove IPNI advertisement relay over pubsub via Lotus node as it now has been deprecated. ([filecoin-project/lotus#12768](https://github.com/filecoin-project/lotus/pull/12768)
+- Add `LOTUS_MIGRATE_PROGRESS_LOG_SECONDS` environment variable to control the log period of the migration progress. ([filecoin-project/lotus#12732](https://github.com/filecoin-project/lotus/pull/12732))
 
 # UNRELEASED v.1.32.0